requirements.txt
google-generativeai
python-dotenv>=0.19.0
aiohttp>=3.8.0
pytest
pytest-asyncio
requests>=2.26.0
flask[async]
asgiref
